requirements.txt
transformers >= 4.30

pii-data >= 0.4.0, <1.0.0
pii-extract-base >= 0.6.1, <1.0.0
